Onprogress

See merge request pantau-group/Client/pantau-bersama-android!67

+++ app/src/main/java/com/pantaubersama/app/base/BaseRecyclerAdapter.kt
@@ -109,7 +109,7 @@ abstract class BaseRecyclerAdapter : RecyclerView.Adapter<RecyclerView.ViewHolde
     fun setLoaded() {
         if (isLoadingMore) {
             isLoadingMore = false
-            deleteItem(itemCount - 1)
+            if (data[itemCount - 1] is LoadingModel) deleteItem(itemCount - 1)
         }
     }
 
@@ -122,6 +122,10 @@ abstract class BaseRecyclerAdapter : RecyclerView.Adapter<RecyclerView.ViewHolde
 
     fun isDataEnd(): Boolean {
         return isDataEnd
+    }
+
+    fun getListData(): MutableList<ItemModel> {
+        return data
     }
 
     override fun getItemViewType(position: Int): Int {
